requirements.txt
fastapi>=0.100.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-docx>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
python-multipart>=0.0.5
pydantic-settings>=2.0.0
